Redesign process completion signaling.
Process completion needs to be signaled by writing to a specific
file. However, this writing needed to be executed using srun,
which is a computationally expensive operation. This was necessary
to avoid those situations in which a job was being cancelled
by SLURM but in spite of this, it was able to signal process
completion, resulting in a corrupted state. In the new code,
this problem is avoided by trapping the SIGTERM function used by
SLURM to notify job cancellation. In addition to this, the sbatch
command uses the -signal option to request SLURM to notify the job
in advance about its cancellation.