Redesign treatment of options.
Options are now written to a single file or to a set of blocks
of a fixed number of lines given by the OPT_FILE_LINES_PER_BLOCK
constant. This allows to avoid the increase in computational cost
when retrieving options for arrays of increasing size. This is
due to the fact that the options for a particular task are
now retrieved from fixed size files.